requirements.txt
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.1
attrs==21.4.0
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
fastapi==0.75.2
flake8==4.0.1
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
Mako==1.2.0
MarkupSafe==2.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pydantic==1.9.0
pyflakes==2.4.0
pyparsing==3.0.8
pytest==7.1.2
python-dotenv==0.20.0
PyYAML==6.0
requests==2.27.1
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.17.1
tomli==2.0.1
typing-extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
